refactor auditer start, update tests
