+++ lib/audit-ci.js
@@ -148,38 +148,13 @@ function getPackageManagerType(pmArg, directory) {
 
 const pm = getPackageManagerType(argv.p, argv.d);
 
-const auditor =
-  pm === 'npm' ? require('./npm-auditer') : require('./yarn-auditer');
-
-const RETRY_ERROR_MSG = {
-  npm: `${
-    argv.registry
-      ? `npm ERR! audit Your configured registry (${argv.registry}) `
-      : ''
-  }does not support audit requests.`,
-  yarn: '503 Service Unavailable',
-};
-
-function audit(attempt = 1) {
-  auditor
-    .audit(argv)
-    .then(() => {
-      console.log('\x1b[32m%s\x1b[0m', `Passed ${pm} security audit.`);
-    })
-    .catch(err => {
-      const message = err.message || err;
-      if (
-        attempt < argv['retry-count'] &&
-        message &&
-        message.includes(RETRY_ERROR_MSG[pm])
-      ) {
-        return audit(attempt + 1);
-      }
-      console.error('\x1b[31m%s\x1b[0m', message);
-      console.error('\x1b[31m%s\x1b[0m', 'Exiting...');
-      process.exitCode = 1;
-      return null;
-    });
-}
-
-audit();
+require('./audit')(pm, argv)
+  .then(() => {
+    console.log('\x1b[32m%s\x1b[0m', `Passed ${pm} security audit.`);
+  })
+  .catch(err => {
+    const message = err.message || err;
+    console.error('\x1b[31m%s\x1b[0m', message);
+    console.error('\x1b[31m%s\x1b[0m', 'Exiting...');
+    process.exitCode = 1;
+  });
