+++ test/npm-auditer.js
@@ -5,7 +5,8 @@
  */
 const { expect } = require('chai');
 const path = require('path');
-const { audit } = require('../lib/npm-auditer');
+// const { audit } = require('../lib/npm-auditer');
+const audit = require('../lib/audit').bind(null, 'npm');
 
 function config(additions) {
   const defaultConfig = {
@@ -19,6 +20,7 @@ function config(additions) {
     advisories: [],
     whitelist: [],
     'show-not-found': false,
+    'retry-count': 5,
     directory: './',
     registry: undefined,
   };
